🐛 fix(types): size numeric fields by their real semantics
- Database.port: u32 -> u16 (TCP ports are 16-bit; matches the std
  convention of SocketAddr::port)
- AppMetrics.net and DatabaseMetrics.net: [u32; 2] -> [u64; 2] (byte
  counters overflow u32 past 4 GiB of traffic in the sampled window;
  every other byte counter in the crate is already u64)
- AnalyticsItem.visits/requests: u32 -> u64 (traffic counters,
  consistent with the u64 request counters in the network types)

+++ src/types/analytics.rs
@@ -8,9 +8,9 @@ pub struct AnalyticsItem {
     #[serde(rename = "type")]
     pub item_type: String,
     /// Total number of unique visits attributed to this item.
-    pub visits: u32,
+    pub visits: u64,
     /// Total number of HTTP requests attributed to this item.
-    pub requests: u32,
+    pub requests: u64,
     /// Total bytes transferred for this item.
     pub bytes: u64,
     /// The time period these metrics cover, as an ISO 8601 date string.
